requirements.txt
absl-py==2.0.0
asgiref==3.7.2
blinker==1.9.0
boto3==1.33.4
botocore==1.33.4
cachetools==5.3.2
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.10.0
Django==4.2.7
faiss-cpu==1.10.0
filelock==3.13.1
fire==0.5.0
Flask==3.1.0
fonttools==4.45.1
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.4
google-auth-oauthlib==1.1.0
grpcio==1.59.3
huggingface-hub==0.19.4
idna==2.10
image==1.5.33
itsdangerous==2.2.0
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.5
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
oauthlib==3.2.2
opencv-python==4.8.1.78
opencv-python-headless==4.8.0.74
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
protobuf==4.23.4
psutil==5.9.6
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybboxes==0.1.6
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
regex==2024.11.6
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
roboflow==1.1.10
rsa==4.9
s3transfer==0.8.2
safetensors==0.5.3
sahi==0.11.15
scipy==1.11.4
seaborn==0.13.0
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sqlparse==0.4.4
supervision==0.16.0
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
termcolor==2.3.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
tokenizers==0.14.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.66.1
transformers==4.35.0
typing_extensions==4.13.1
tzdata==2023.3
urllib3==2.0.7
Werkzeug==3.0.1
yolov5==7.0.13
